# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kira Code API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/organizations": {
      "get": {
        "tags": [
          "organizations"
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse42"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "organizations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse43"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{id}": {
      "get": {
        "tags": [
          "organizations"
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse44"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "organizations"
        ]
      },
      "patch": {
        "tags": [
          "organizations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse45"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{org_id}/invitations": {
      "get": {
        "tags": [
          "organizations"
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse47"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "organizations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse46"
                }
              }
            }
          }
        }
      }
    },
    "/organizations/{org_id}/invitations/revoke": {
      "post": {
        "tags": [
          "organizations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeInvitationRequest"
              }
            }
          },
          "required": true
        }
      }
    },
...
    "/organizations/{org_id}/members/{user_id}/role": {
      "patch": {
        "tags": [
          "organizations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse51"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse42": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListOrganizationsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListOrganizationsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "ListOrganizationsResponse": {
        "type": "object",
        "properties": {
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationWithRole"
            }
          }
        },
        "required": [
          "organizations"
        ]
      },
      "OrganizationWithRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "is_personal": {
            "type": "boolean"
          },
          "issue_prefix": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_role": {
            "$ref": "#/components/schemas/MemberRole"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "is_personal",
          "issue_prefix",
          "created_at",
          "updated_at",
          "user_role"
        ]
      },
      "MemberRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "MEMBER"
        ]
      },
      "CreateOrganizationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "ApiResponse43": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateOrganizationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateOrganizationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "CreateOrganizationResponse": {
        "type": "object",
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationWithRole"
          }
        },
        "required": [
          "organization"
        ]
      },
      "ApiResponse44": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GetOrganizationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GetOrganizationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "GetOrganizationResponse": {
        "type": "object",
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "user_role": {
            "type": "string"
          }
        },
        "required": [
          "organization",
          "user_role"
        ]
      },
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "is_personal": {
            "type": "boolean"
          },
          "issue_prefix": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "is_personal",
          "issue_prefix",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateOrganizationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "ApiResponse45": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organization"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organization"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "CreateInvitationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "ApiResponse46": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateInvitationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateInvitationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "CreateInvitationResponse": {
        "type": "object",
        "properties": {
          "invitation": {
            "$ref": "#/components/schemas/Invitation"
          }
        },
        "required": [
          "invitation"
        ]
      },
      "Invitation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "invited_by_user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          },
          "status": {
            "$ref": "#/components/schemas/InvitationStatus"
          },
          "token": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organization_id",
          "email",
          "role",
          "status",
          "token",
          "created_at",
          "expires_at"
        ]
      },
      "InvitationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "DECLINED",
          "EXPIRED"
        ]
      },
      "ApiResponse47": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListInvitationsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListInvitationsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "ListInvitationsResponse": {
        "type": "object",
        "properties": {
          "invitations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invitation"
            }
          }
        },
        "required": [
          "invitations"
        ]
      },
      "RevokeInvitationRequest": {
        "type": "object",
        "properties": {
          "invitation_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "invitation_id"
        ]
      },
...
      "UpdateMemberRoleRequest": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          }
        },
        "required": [
          "role"
        ]
      },
      "ApiResponse51": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UpdateMemberRoleResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UpdateMemberRoleResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "UpdateMemberRoleResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          }
        },
        "required": [
          "user_id",
          "role"
        ]
      },
...
    }
  }
}